pyproject.toml
[project]
name = "slink"
version = "0.1.0"
description = ""
authors = [
    {name = "inksne",email = "inksne@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "slink", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.13"
namespace_packages = true
check_untyped_defs = true
show_error_codes = true
strict = true
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "sqlalchemy.*",
    "asyncpg.*"
]
ignore_errors = true
ignore_missing_imports = true